// ui/packages/src/api/generated/filesys/schemas/storagePlatformDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Refactored free-fs with a subject-match-java style layered architecture.
 * OpenAPI spec version: v1.0.0
 */

export interface StoragePlatformDTO {
  /** 配置模式（JSON格式） */
  configSchema?: string;
  /** 描述信息 */
  description?: string;
  /** 图标 */
  icon?: string;
  /** 平台ID */
  id?: number;
  /** 平台标识符 */
  identifier?: string;
  /** 是否为默认平台 */
  isDefault?: number;
  /** 链接地址 */
  link?: string;
  /** 平台名称 */
  name?: string;
}